+++ src/port_diff/rewrite.rs
@@ -10,7 +10,7 @@ use crate::{
     Graph, PortDiff,
 };
 
-use super::{EdgeData, Owned, PortDiffData};
+use super::{BoundaryPort, EdgeData, Owned, PortDiffData};
 
 #[derive(Error, Debug)]
 pub enum InvalidRewriteError {
@@ -29,12 +29,12 @@ impl<G: Graph> PortDiff<G> {
     /// The function `boundary_map` will be called once for every boundary port
     /// of the new diff. It is passed as argument an owned port, the image of
     /// the boundary port in a parent diff. It must return the site of the
-    /// boundary port in the new graph.
+    /// boundary port in the new graph, or a sentinel node.
     pub fn rewrite(
         nodes: impl IntoIterator<Item = Owned<G::Node, G>>,
         edges: impl IntoIterator<Item = (Owned<Port<G>, G>, Owned<Port<G>, G>)>,
         new_graph: G,
-        mut boundary_map: impl FnMut(Owned<Port<G>, G>) -> Site<G::Node, G::PortLabel>,
+        mut boundary_map: impl FnMut(Owned<Port<G>, G>) -> BoundaryPort<G>,
     ) -> Result<Self, InvalidRewriteError> {
         // Collect nodes per portdiff
         let nodes: BTreeMap<_, BTreeSet<_>> =
@@ -137,7 +137,11 @@ impl<G: Graph> PortDiff<G> {
                 }
             }
             for b in diff.boundary_iter() {
-                if !subgraph.nodes().contains(&diff.boundary_site(b).node) {
+                let Some(site) = diff.boundary_site(b) else {
+                    // Sentinel boundaries cannot be rewritten
+                    continue;
+                };
+                if !subgraph.nodes().contains(&site.node) {
                     continue;
                 }
                 if !used_unbound_ports.remove(&b) {
@@ -185,14 +189,14 @@ impl<G: Graph> PortDiff<G> {
     pub fn rewrite_edges(
         edges: impl IntoIterator<Item = (Owned<Port<G>, G>, Owned<Port<G>, G>)> + Clone,
         new_graph: G,
-        boundary_map: impl FnMut(Owned<Port<G>, G>) -> Site<G::Node, G::PortLabel>,
+        boundary_map: impl FnMut(Owned<Port<G>, G>) -> BoundaryPort<G>,
     ) -> Result<Self, InvalidRewriteError> {
         let nodes: BTreeSet<_> = edges
             .clone()
             .into_iter()
             .flat_map(|(l, r)| {
                 [l, r].map(|p| Owned {
-                    data: p.site().node,
+                    data: p.site().unwrap().node, // TODO: what to do with sentinels?
                     owner: p.owner,
                 })
             })
@@ -207,7 +211,7 @@ impl<G: Graph> PortDiff<G> {
         &self,
         nodes: &BTreeSet<G::Node>,
         new_graph: G,
-        mut boundary_map: impl FnMut(Port<G>) -> Site<G::Node, G::PortLabel>,
+        mut boundary_map: impl FnMut(Port<G>) -> BoundaryPort<G>,
     ) -> Result<Self, InvalidRewriteError> {
         let edges = self
             .graph()
@@ -407,11 +411,12 @@ mod tests {
             let n1 = rhs.add_node(1, 0);
             rhs.link_nodes(n0, 3, n1, 0).unwrap();
             parent.rewrite_induced(&BTreeSet::from_iter([v]), rhs, |p| {
-                let offset = Owned::new(p, parent.clone()).site().port;
+                let offset = Owned::new(p, parent.clone()).site().unwrap().port;
                 Site {
                     node: n0,
                     port: offset,
                 }
+                .into()
             })
         };
         let (_, n1, n2, _) = PortView::nodes_iter(&parent.graph).collect_tuple().unwrap();
@@ -432,11 +437,11 @@ mod tests {
 
         let child_a_out0 = child_a
             .boundary_iter()
-            .find(|&bd| child_a.boundary_site(bd).port == PortOffset::Outgoing(0))
+            .find(|&bd| child_a.boundary_site(bd).unwrap().port == PortOffset::Outgoing(0))
             .unwrap();
         let child_b_in0 = child_b
             .boundary_iter()
-            .find(|&bd| child_b.boundary_site(bd).port == PortOffset::Incoming(0))
+            .find(|&bd| child_b.boundary_site(bd).unwrap().port == PortOffset::Incoming(0))
             .unwrap();
         let cross_edge = (
             Owned::new(Port::Boundary(child_a_out0), child_a.clone()),
@@ -451,13 +456,15 @@ mod tests {
             if n.owner == child_a {
                 Site {
                     node: n0,
-                    port: n.site().port,
+                    port: n.site().unwrap().port,
                 }
+                .into()
             } else {
                 Site {
                     node: n1,
-                    port: n.site().port,
+                    port: n.site().unwrap().port,
                 }
+                .into()
             }
         })
         .unwrap();